ngx_http_ubus_module: make thread work directly on res string
Make thread directly work on the result string instead of referencing
them with an index. This is to improve data handling by making
everything self contained instead of having to access data in the
request context struct.

Object call is updated to also allign to this res_str implementation by
refereincing a local variable of it.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>