ngx_http_ubus_module: rework thread sched logic for array processing
Rework thread sheduling logic for array processing. The current logic
was flawed and very slow since it does wait for the other thread to
finish before looping again and fire the other one.

Rework this and make the thread stateless, by using an alternative way
to signal thread finish. With the new implementation we don't need to
join each thread but we just check if every object has been cosumed and
processed.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>