ngx_http_ubus_module: fix memory leak in ubus_post_object
Fix memory leak in ubus_post_object. Since we were using ctx->buf, and
we later reuse ctx-buf to store the final string, we were losing the
pointer to the blob_buf allocated in ubus_post_object to parse the
jsobj. Drop usage of ctx->buf in ubuf_post_object and handle it locally
in the internal function.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>