ngx_http_ubus_module: rework to nginx thread implementation
From nginx developer guide:

It is recommended to avoid using threads in nginx because it will definitely break things: most nginx functions are not thread-safe. It is expected that a thread will be executing only system calls and thread-safe library functions. If you need to run some code that is not related to client request processing, the proper way is to schedule a timer in the init_process module handler and perform required actions in timer handler. Internally nginx makes use of threads to boost IO-related operations, but this is a special case with a lot of limitations.

New version of nginx provide an internal implementation for threads
using thread pool logic. Convert the module to this new implementation.

Major rework were needed to make everything non blocking (or half
blocking with the use of semaphores)

The completition handler is called only AFTER the request is completed
hence a complete new logic needs to be applied. The old implementation
with spawining a thread for each thread is reworked to follow the new
nginx thread api.

On top of this, an additional thread is needed to complete the request
by checking the obj_complete semaphores and append the entire data and
send the header, send the body and finalize the request.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>