Revert "ngx_http_ubus_module: rework ubus connection to per thread"
This reverts commit 367fe594261dff19083312858f6423e05d28128e.

Ubus is actually very problematic with concurrent connection. It was
notice that many problems and deadlock happens in the very rare (but
actually not that rare due to all the thread implementation and a
webserver handling multiple connection at the same time) and this also
seems to cause heap corruption.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>