ngx_http_ubus_module: make thread support optional
Make nginx thread support optional. The module is still very suggested
to run with thread support and thread pool but can also be compiled
without. The module is implemented with the assumption of thread support
and without it the task functions are adapted to use the single thread
version.

With single thread implementation the global ubus mutex is not needed
hence the lock/unlock call are wrapped in a MACRO and compiled out if
thread support is not supported.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>